+++ api-base/src/commons/jobs/exampleJob.ts
@@ -1,0 +1,8 @@
+import { Job, ZanixJob } from '@zanix/server'
+
+@Job('jobExample')
+export class ExampleJob extends ZanixJob {
+  public onProcess(secureData: string) {
+    this.logger.debug(`Job was executed. Data: ${secureData}`)
+  }
+}