+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Example tasker and job
+- New server version on `api-base` template (`@zanix/server@1.1.0`)
+
 ## [1.0.2] - 2023-06-15
 
 ### Added